Replace Reachability framework with one that correctly determines network status #99
I added a hello endpoint on our new backend that just returns the word "hello". When the reachability class determines the network is available, we poll that address immediately to see if we get the correct response. If it succeeds, we send the network available notification to begin syncing. If it fails, we retry every second until it either succeeds or the reachability class determines the network is unavailable.

This should fix the issue whereby the network interface is ready, but network connectivity isn't, causing initial syncs to fail.

Also I moved the Reachability.swift file into our project instead of using it as a Carthage framework since it's literally 1 file.